Cadastro e listagem de produtos funcionando

+++ src/cadastroVIEW.java
@@ -149,7 +149,10 @@ public class cadastroVIEW extends javax.swing.JFrame {
         String status = "A Venda";
         
         produto.setNome(nome);
-        produto.setValor(Integer.parseInt(valor));
+        
+        
+        produto.setValor(Double.parseDouble(valor));
+        
         produto.setStatus(status);
 
         ProdutosDAO produtodao = new ProdutosDAO();
